API

KM3io.AcousticSignal

KM3io.AcousticsTriggerParameter

KM3io.Detector

KM3io.Detector

KM3io.Detector

KM3io.DetectorModule

KM3io.Hydrophone

KM3io.Location

KM3io.PMT

KM3io.Tripod

KM3io.Waveform

Base.getindex

Base.read

Base.read

Base.read

Base.read

Base.read

Base.write

Base.write

KM3io._extract_comments

KM3io.calibrate

KM3io.floordist

KM3io.hydrophoneenabled

KM3io.nthbitset

KM3io.piezoenabled

KM3io.slew

KM3io.AcousticSignal — Type

AcousticSignal is a custom type with four fields to store all the information inside the raw acoustic binary files.

dom_id::Int32 ID of the module

utc_seconds:: UInt32 storing the first 4 Bytes and is a UNIX time stamp

ns_cycles:: UInt32 storing the second 4 Bytes

samples:: UInt32 storing the third 4 Bytes, corresponding to the number of data points accuired during the measring window

pcm:: Vector of Float32 of length frame_length, storing all other 4 Byte blocks. Each entry is a data point of the acoustic signal.

KM3io.AcousticsTriggerParameter — Type

Certain parameters which define an acoustic event.

KM3io.Detector — Type

A KM3NeT detector.

KM3io.Detector — Method

function Detector(filename::AbstractString)

Create a Detector instance from a DETX file.

KM3io.Detector — Method

function Detector(io::IO)

Create a Detector instance from an IO stream.

KM3io.DetectorModule — Type

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

KM3io.Hydrophone — Type

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

KM3io.Location — Type

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

KM3io.PMT — Type

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

KM3io.Tripod — Type

A tripod installed on the seabed which sends acoustic signals to modules.

KM3io.Waveform — Type

Waveform translates Emitter ID to Tripod ID.

Base.getindex — Method

Base.getindex(d::DetectorModule, i) = d.pmts[i+1]

The index in this context is the DAQ channel ID of the PMT, which is counting from 0.

Base.read — Function

function read(filename::AbstractString,T::Type{AcousticSignal}, overlap::Int=DAQ_ADF_ANALYSIS_WINDOW_OVERLAP)

Reads in a raw binary acoustics file.

Base.read — Method

function read(filename::AbstractString, T::Type{AcousticsTriggerParameter})

Reads the 'acousticstriggerparameters.txt' file.

Base.read — Method

function read(filename::AbstractString, T::Type{Hydrophone})

Reads a vector of Hydrophones from an ASCII file.

Base.read — Method

function read(filename:AbstractString, T::Type{Tripod})

Reads a vector of Tripods from an ASCII file.

Base.read — Method

function read(filename::AbstractString, T::Type{Waveform})

Reads the waveform ASCII file.

Base.write — Method

function write(filename::AbstractString, tripods::Dict{Int8, Tripod})

Writes the position of tripods out into an ASCII file.

Base.write — Method

function write(io::IO, d::Detector; version=:same)

Writes the detector to a DETX formatted file. The target version can be specified via the version keyword. Note that if converting to higher versions, missing parameters will be filled with reasonable default values. In case of downgrading, information will be lost.

KM3io._extract_comments — Method

function _extract_comments(lines<:Vector{AbstractString}, prefix<:AbstractString)

Returns only the lines which are comments, identified by the prefix. The prefix is omitted.

KM3io.calibrate — Method

calibrate(det, hits)

Apply geometry and time calibration to given hits.

KM3io.floordist — Method

function floordist(calib::Calibration)

Calculates the average floor distance between neighboured modules.

KM3io.hydrophoneenabled — Method

function hydrophonenabled(m::DetectorModule)

Return true if the hydrophone is enabled, false otherwise.

KM3io.nthbitset — Method

nthbitset(n, a) = !Bool((a >> (n - 1)) & 1)

Return true if the n-th bit of a is set, false otherwise.

KM3io.piezoenabled — Method

function piezoenabled(m::DetectorModule)

Return true if the piezo is enabled, false otherwise.

KM3io.slew — Method

slew(tot)

Return the time slewing for a ToT.