API

KM3io.Detector

KM3io.DetectorModule

KM3io.Direction

KM3io.Hydrophone

KM3io.Location

KM3io.PMT

KM3io.Position

KM3io.Tripod

KM3io.Waveform

Basic Data Structures

KM3io.Position — Type

A vector to represent a position in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Direction — Type

A vector to represent a direction in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Location — Type

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

Fields

string::Int32

floor::Int8

Hardware Components

KM3io.PMT — Type

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position

dir::Direction

t₀::Float64

status::Union{Missing, Int32}

KM3io.DetectorModule — Type

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position

location::Location

n_pmts::Int8

pmts::Vector{PMT}

q::Union{Missing, Quaternion}

status::Int32

t₀::Float64

KM3io.Detector — Type

A KM3NeT detector.

Fields

version::Int8

id::Int32

validity::Union{Missing, KM3io.DateRange}

pos::Union{Missing, UTMPosition}

utm_ref_grid::Union{Missing, String}

n_modules::Int32

modules::Dict{Int32, DetectorModule}

locations::Dict{Tuple{Int64, Int64}, DetectorModule}

strings::Vector{Int64}

comments::Vector{String}

Acoustics

KM3io.Waveform — Type

Waveform translates Emitter ID to Tripod ID.

Fields

ids::Dict{Int8, Int8}

Missing docstring.

Missing docstring for AcousticSignal. Check Documenter's build log for details.

KM3io.Hydrophone — Type

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

Fields

location::Location

pos::Position

KM3io.Tripod — Type

A tripod installed on the seabed which sends acoustic signals to modules.

Fields

id::Int8

pos::Position