API

KM3io.AcousticSignal

KM3io.AcousticsTriggerParameter

KM3io.CHClient

KM3io.CalibratedHit

KM3io.CalibratedMCHit

KM3io.CherenkovPhoton

KM3io.Detector

KM3io.DetectorModule

KM3io.Direction

KM3io.Evt

KM3io.H5CompoundDataset

KM3io.H5File

KM3io.Hydrophone

KM3io.K40Rates

KM3io.Location

KM3io.MCTrk

KM3io.PMT

KM3io.Position

KM3io.RecStageRange

KM3io.Tripod

KM3io.Trk

KM3io.Waveform

KM3io.azimuth

KM3io.bestaashower

KM3io.bestjppmuon

KM3io.bestjppshower

KM3io.besttrack

KM3io.calibrate

KM3io.categorize

KM3io.center

KM3io.cherenkov

KM3io.floordist

KM3io.hashistory

KM3io.hydrophoneenabled

KM3io.is3dmuon

KM3io.is3dshower

KM3io.ismxshower

KM3io.isnb

KM3io.most_frequent

KM3io.nthbitset

KM3io.phi

KM3io.piezoenabled

KM3io.slew

KM3io.theta

KM3io.triggered

KM3io.zenith

Basic Data Structures

KM3io.Position — Type

struct Position{T} <: StaticArraysCore.FieldVector{3, T}

A vector to represent a position in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Direction — Type

struct Direction{T<:AbstractFloat} <: StaticArraysCore.FieldVector{3, T<:AbstractFloat}

A vector to represent a direction in 3D.

Fields

x::AbstractFloat

y::AbstractFloat

z::AbstractFloat

KM3io.Location — Type

struct Location

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

Fields

string::Int32

floor::Int8

Offline Format

KM3io.Evt — Type

struct Evt

An offline event.

Fields

id::Int64

det_id::Int64

mc_id::Int64

run_id::Int64

mc_run_id::Int64

frame_index::Int64

trigger_mask::UInt64

trigger_counter::UInt64

overlays::UInt64

t::KM3io.UTCTime

hits::Vector{CalibratedHit}

trks::Vector{Trk}

w::Vector{Float64}

w2list::Vector{Float64}

w3list::Vector{Float64}

mc_event_time::KM3io.UTCTime

mc_t::Float64

mc_hits::Vector{CalibratedMCHit}

mc_trks::Vector{MCTrk}

index::Int64

flags::Int64

usr::Dict{String, Float64}

KM3io.CalibratedHit — Type

struct CalibratedHit <: AbstractCalibratedHit

A calibrated hit of the offline dataformat. Caveat: the position, direction and t fields might still be 0 due to the design philosophy of the offline format (one class for all).

Fields

dom_id::Int32

channel_id::UInt32

tdc::UInt32

tot::UInt32

trigger_mask::UInt64

t::Float64

pos::Position{Float64}

dir::Direction{Float64}

KM3io.CalibratedMCHit — Type

struct CalibratedMCHit

A calibrated MC hit of the offline dataformat. Caveat: the position and direction fields might still be 0 due to the design philosophy of the offline format (one class for all).

Fields

pmt_id::Int32

t::Float64

a::Float64

type::Int32

origin::Int32

pos::Position{Float64}

dir::Direction{Float64}

KM3io.Trk — Type

struct Trk

Represents a reconstructed "track", which can be e.g. a muon track but also a shower.

Fields

id::Int64

pos::Position{Float64}

dir::Direction{Float64}

t::Float64

E::Float64

len::Float64

lik::Float64

rec_type::Int32

rec_stages::Vector{Int32}

fitinf::Vector{Float64}

KM3io.MCTrk — Type

struct MCTrk

A simulated (Monte Carlo, hence "MC") track (or shower).

Fields

id::Int64

pos::Position{Float64}

dir::Direction{Float64}

t::Float64

E::Float64

len::Float64

type::Int32

status::Int32

mother_id::Int32

counter::Int32

Online Format

HDF5

KM3io.H5File — Type

struct H5File

A wrapper for an HDF5 file used in KM3NeT.

Fields

_h5f::HDF5.File

_datasets::Dict{String, H5CompoundDataset}

KM3io.H5CompoundDataset — Type

struct H5CompoundDataset{T}

A flat HDF5 compound dataset which is essentially a vector of structs. It has a cache which is filled when elements are pushed to it. The cache is automatically written to the target HDF5 path when full.

Fields

dset::HDF5.Dataset

cache::KM3io.H5CompoundDatasetCache

Hardware

KM3io.PMT — Type

struct PMT

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

dir::Direction{Float64}

t₀::Float64

status::Union{Missing, Int32}

KM3io.DetectorModule — Type

struct DetectorModule

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

location::Location

n_pmts::Int8

pmts::Vector{PMT}

q::Union{Missing, Quaternion{Float64}}

status::Int32

t₀::Float64

KM3io.Detector — Type

struct Detector

A KM3NeT detector.

Fields

version::Int8

id::Int32

validity::Union{Missing, KM3io.DateRange}

pos::Union{Missing, UTMPosition{Float64}}

utm_ref_grid::Union{Missing, String}

n_modules::Int32

modules::Dict{Int32, DetectorModule}

locations::Dict{Tuple{Int64, Int64}, DetectorModule}

strings::Vector{Int64}

comments::Vector{String}

KM3io.Hydrophone — Type

struct Hydrophone

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

Fields

location::Location

pos::Position{Float64}

KM3io.Tripod — Type

struct Tripod

A tripod installed on the seabed which sends acoustic signals to modules.

Fields

id::Int8

pos::Position{Float64}

KM3io.piezoenabled — Function

piezoenabled(m::DetectorModule) -> Bool

function piezoenabled(m::DetectorModule)

Return true if the piezo is enabled, false otherwise.

KM3io.hydrophoneenabled — Function

hydrophoneenabled(m::DetectorModule) -> Bool

function hydrophonenabled(m::DetectorModule)

Return true if the hydrophone is enabled, false otherwise.

KM3io.center — Function

center(m::DetectorModule) -> Position{Float64}

Calculate the centre of a module by fitting the crossing point of the PMT axes.

center(d::Detector) -> Position{Float64}

Calculate the center of the detector based on the location of the optical modules.

Optical Data

Acoustics

KM3io.Waveform — Type

struct Waveform

Waveform translates Emitter ID to Tripod ID.

Fields

ids::Dict{Int8, Int8}

KM3io.AcousticSignal — Type

struct AcousticSignal

AcousticSignal is a custom type with four fields to store all the information inside the raw acoustic binary files.

dom_id::Int32 ID of the module

utc_seconds:: UInt32 storing the first 4 Bytes and is a UNIX time stamp

ns_cycles:: UInt32 storing the second 4 Bytes

samples:: UInt32 storing the third 4 Bytes, corresponding to the number of data points accuired during the measring window

pcm:: Vector of Float32 of length frame_length, storing all other 4 Byte blocks. Each entry is a data point of the acoustic signal.

Fields

dom_id::Int32

utc_seconds::UInt32

ns_cycles::UInt32

samples::UInt32

pcm::Vector{Float32}

KM3io.AcousticsTriggerParameter — Type

struct AcousticsTriggerParameter

Certain parameters which define an acoustic event.

Fields

q::Float64

tmax::Float64

nmin::Int32

Calibration

KM3io.calibrate — Function

calibrate(det::Detector, hits) -> Vector{XCalibratedHit}

Apply geometry and time calibration to given hits.

KM3io.floordist — Function

floordist(det::Detector) -> Float64

Calculates the average floor distance between neighboured modules.

KM3io.slew — Function

slew(tot::Integer) -> Any

Return the time slewing for a ToT.

Physics

KM3io.azimuth — Function

azimuth(d::Direction) -> Any

Calculates the azimuth from a given Direction according to KM3NeT conventions.

KM3io.zenith — Function

zenith(d::Direction) -> Any

Calculates the zenith from a given Direction according to KM3NeT conventions.

KM3io.phi — Function

phi(d::Direction) -> Any

Calculate the phi angle for a given direction

KM3io.theta — Function

theta(d::Direction) -> Any

Calculate the theta angle for a given direction

KM3io.cherenkov — Function

cherenkov(
    track,
    hits::Array{T<:AbstractCalibratedHit, 1}
) -> Any

Calculates the parameters of cherenkov photons emitted from a track and hitting the PMTs represented as (calibrated) hits. The returned cherenkov photons hold information about the closest distance to track, the time residual, arrival time, impact angle, photon travel distance, track travel distance and photon travel direction. See CherenkovPhoton for more information.

KM3io.CherenkovPhoton — Type

struct CherenkovPhoton

A Cherenkov photon with parameters calculated from its inducing track. See cherenkov() for more information.

Fields

d_closest::Float64

d_photon::Float64

d_track::Float64

t::Float64

Δt::Float64

impact_angle::Float64

dir::Direction{Float64}

KM3io.K40Rates — Type

struct K40Rates

K40 rates with L0 and higher level rates (with increasing multiplicities).

Fields

L0::Float64

L1::Vector{Float64}

Trigger

KM3io.triggered — Function

triggered(e) -> Vector

Return true if the passed object (hit, event, ...) was triggered by any trigger algorithm.

KM3io.is3dmuon — Function

is3dmuon(e::DAQEvent) -> Bool

is3dmuon(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:12.

Return true the 3D Muon trigger bit is set.

KM3io.is3dshower — Function

is3dshower(e::DAQEvent) -> Bool

is3dshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:20.

Return true if the 3D Shower trigger bit is set.

KM3io.ismxshower — Function

ismxshower(x) -> Bool

ismxshower(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:28.

ismxshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:29.

Return true if the MX Shower trigger bit is set.

KM3io.isnb — Function

isnb(x) -> Bool

isnb(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:36.

isnb(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:37.

Return true if the NanoBeacon trigger bit is set.

ControlHost

KM3io.CHClient — Type

struct CHClient{T}

A ControlHost client which can communicate with a Ligier dispatcher to receive messages for all the subscribed tags.

To connect to a Ligier which is receiving triggered DAQ events e.g. in the KM3NeT monitoring system or in a test setup consisting of a JLigier dispatcher and a JRegurgitate instance which is redispatching DAQ events (JDAQEvent) from a ROOT file in online format to the JLigier, a CHClient can be created to subscribe the event messages with

julia> using KM3io

julia> c = CHClient{DAQEvent}(ip"127.0.0.1", 5553)

julia> for event in c
           @show event
       end
e = KM3io.DAQEvent with 127 snapshot and 6 triggered hits
e = KM3io.DAQEvent with 147 snapshot and 6 triggered hits
e = KM3io.DAQEvent with 154 snapshot and 8 triggered hits
e = KM3io.DAQEvent with 152 snapshot and 6 triggered hits
...
...
...

Fields

ip::Sockets.IPv4

port::UInt16

tags::Vector{CHTag}

socket::Sockets.TCPSocket

Tools

General tools

KM3io.categorize — Function

categorize(field::Symbol, elements::Vector) -> Dict

Categorise the struct elements of a vector by a given field into a dictionary of T.field => Vector{T}.

Examples

julia> using KM3io

julia> struct PMT  # just an ad-hoc PMT struct for demonstration purposes
         dom_id
         time
       end

julia> pmts = [PMT(2, 10.4), PMT(4, 23.5), PMT(2, 42.0)];

julia> categorize(:dom_id, pmts)
Dict{Any, Vector{PMT}} with 2 entries:
  4 => [PMT(4, 23.5)]
  2 => [PMT(2, 10.4), PMT(2, 42.0)]

KM3io.nthbitset — Function

nthbitset(n, a) -> Any

Return true if the n-th bit of a is set, false otherwise.

KM3io.most_frequent — Function

most_frequent(iterable) -> Any

Return the most frequent value of a given iterable.

most_frequent(f::Function, iterable; rettype) -> Int64

Return the most frequent value of a given iterable based on the return value of a function f which returns (hashable) values of rettype.

Reconstruction

KM3io.besttrack — Function

besttrack(
    e::Evt,
    rec_type::Integer,
    rsr::RecStageRange
) -> Union{Nothing, Trk}

Return the best reconstructed track for a given reconstruction type and reconstruction stage range. If no track could be found, nothing is returned.

KM3io.bestjppmuon — Function

bestjppmuon(e::Evt) -> Union{Nothing, Trk}

Returns the best reconstructed JMuon track of an event or nothing if there are none.

bestjppmuon(trks::Vector{Trk}) -> Union{Nothing, Trk}

Returns the best reconstructed JMuon track or nothing if there are none.

KM3io.bestjppshower — Function

bestjppshower(e::Evt) -> Union{Nothing, Trk}

Returns the best reconstructed JShower "track" of an event or nothing if there are none.

bestjppshower(trks::Vector{Trk}) -> Union{Nothing, Trk}

Returns the best reconstructed JShower "track" or nothing if there are none.

KM3io.bestaashower — Function

bestaashower(e::Evt) -> Union{Nothing, Trk}

Returns the best reconstructed aashower "track" of an event or nothing if there are none.

bestaashower(trks::Vector{Trk}) -> Union{Nothing, Trk}

Returns the best reconstructed aashower "track" or nothing if there are none.

KM3io.RecStageRange — Type

struct RecStageRange{T<:Integer}

This struct is used to represent a range of reconstruction stages. These are well-defined integers (see KM3NeT Dataformat) for each reconstruction algorithm and are stored in a vector named rec_stages of each Trk.

julia> using KM3io

julia> rsr = RecStageRange(KM3io.RECONSTRUCTION.JMUONBEGIN, KM3io.RECONSTRUCTION.JMUONEND)
RecStageRange{Int64}(0, 99)

julia> KM3io.RECONSTRUCTION.JMUONSIMPLEX ∈ rsr
true

julia> KM3io.RECONSTRUCTION.AASHOWERFITPREFIT ∈ rsr
false

julia> 23 ∈ rsr
true

julia> 523 ∈ rsr
false

Fields

lower::Integer

upper::Integer

KM3io.hashistory — Function

hashistory(
    t::Trk,
    rec_type::Integer,
    rsr::RecStageRange
) -> Bool

Returns true if a track with a given rec_type contains all the reconstruction stages in rsr::RecStageRange.

hashistory(
    t::Trk,
    rec_type::Integer,
    rec_stage::Integer
) -> Union{Missing, Bool}

Returns true if a track with a given rec_type contains the rec_stage.