API

KM3io.AcousticSignal

KM3io.AcousticsTriggerParameter

KM3io.CHClient

KM3io.CalibratedHit

KM3io.CalibratedMCHit

KM3io.CalibratedSnapshotHit

KM3io.CalibratedTriggeredHit

KM3io.CherenkovPhoton

KM3io.Compass

KM3io.DAQEvent

KM3io.Detector

KM3io.DetectorModule

KM3io.Direction

KM3io.EventHeader

KM3io.Evt

KM3io.FitInformation

KM3io.H5CompoundDataset

KM3io.H5File

KM3io.Hydrophone

KM3io.K40Rates

KM3io.Location

KM3io.MCTrk

KM3io.Orientations

KM3io.PMT

KM3io.PMTData

KM3io.PMTFile

KM3io.Position

KM3io.Quaternion

KM3io.RecStageRange

KM3io.SnapshotHit

KM3io.StringMechanics

KM3io.StringMechanicsParameters

KM3io.SummaryFrame

KM3io.Summaryslice

KM3io.SummarysliceHeader

KM3io.SummarysliceIntervalIterator

KM3io.TriggeredHit

KM3io.Tripod

KM3io.Trk

KM3io.UTCExtended

KM3io.UTCTime

KM3io.Waveform

KM3io.XCalibratedHit

Base.angle

Base.flush

Base.read

Base.read

Base.read

Base.read

Base.read

Base.write

Base.write

Base.write

HDF5.create_dataset

KM3io.addmeta

KM3io.azimuth

KM3io.bestaashower

KM3io.bestjppmuon

KM3io.bestjppshower

KM3io.besttrack

KM3io.calibrate

KM3io.calibratetime

KM3io.categorize

KM3io.center

KM3io.cherenkov

KM3io.combine

KM3io.count_active_channels

KM3io.count_fifostatus

KM3io.count_hrvstatus

KM3io.distance

KM3io.fifostatus

KM3io.floordist

KM3io.getevent

KM3io.getmodule

KM3io.getpmt

KM3io.hashistory

KM3io.haslocation

KM3io.hasudptrailer

KM3io.hrvstatus

KM3io.hydrophoneenabled

KM3io.is3dmuon

KM3io.is3dshower

KM3io.isbasemodule

KM3io.ismxshower

KM3io.isnb

KM3io.isopticalmodule

KM3io.maximal_udp_sequence_number

KM3io.modules

KM3io.most_frequent

KM3io.nthbitset

KM3io.number_of_udp_packets_received

KM3io.phi

KM3io.piezoenabled

KM3io.pmtrate

KM3io.pmtrates

KM3io.slerp

KM3io.slew

KM3io.status

KM3io.tdcstatus

KM3io.theta

KM3io.triggered

KM3io.wrstatus

KM3io.zenith

Basic Data Structures

KM3io.Position — Type

struct Position{T} <: StaticArraysCore.FieldVector{3, T}

A vector to represent a position in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Direction — Type

struct Direction{T<:AbstractFloat} <: StaticArraysCore.FieldVector{3, T<:AbstractFloat}

A vector to represent a direction in 3D.

Fields

x::AbstractFloat

y::AbstractFloat

z::AbstractFloat

KM3io.Location — Type

struct Location

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

Fields

string::Int32

floor::Int8

Offline Format

KM3io.Evt — Type

struct Evt

An offline event.

Fields

id::Int64

det_id::Int64

mc_id::Int64

run_id::Int64

mc_run_id::Int64

frame_index::Int64

trigger_mask::UInt64

trigger_counter::UInt64

overlays::UInt64

t::UTCTime

hits::Vector{CalibratedHit}

trks::Vector{Trk}

w::Vector{Float64}

w2list::Vector{Float64}

w3list::Vector{Float64}

mc_event_time::UTCTime

mc_t::Float64

mc_hits::Vector{CalibratedMCHit}

mc_trks::Vector{MCTrk}

index::Int64

flags::Int64

usr::Dict{String, Float64}

KM3io.CalibratedHit — Type

struct CalibratedHit <: AbstractCalibratedHit

A calibrated hit of the offline dataformat. Caveat: the position, direction and t fields might still be 0 due to the design philosophy of the offline format (one class for all).

Fields

dom_id::Int32

channel_id::UInt32

tdc::UInt32

tot::UInt32

trigger_mask::UInt64

t::Float64

pos::Position{Float64}

dir::Direction{Float64}

KM3io.CalibratedMCHit — Type

struct CalibratedMCHit

A calibrated MC hit of the offline dataformat. Caveat: the position and direction fields might still be 0 due to the design philosophy of the offline format (one class for all).

Fields

pmt_id::Int32

t::Float64

a::Float64

type::Int32

origin::Int32

pos::Position{Float64}

dir::Direction{Float64}

KM3io.XCalibratedHit — Type

struct XCalibratedHit <: AbstractCalibratedHit

A fully dressed hit with all calibration information which can be obtained. This structure is similar to the Hit structure in aanet and should be used wisely. Most of the time it's much more performant to use dedicated (simplified) structures.

Fields

dom_id::UInt32

channel_id::UInt32

t::Float64

tot::UInt8

trigger_mask::UInt64

pos::Position{Float64}

dir::Direction{Float64}

t0::Float64

string::UInt8

floor::UInt8

KM3io.Trk — Type

struct Trk

Represents a reconstructed "track", which can be e.g. a muon track but also a shower.

Fields

id::Int64

pos::Position{Float64}

dir::Direction{Float64}

t::Float64

E::Float64

len::Float64

lik::Float64

rec_type::Int32

rec_stages::Vector{Int32}

fitinf::FitInformation

KM3io.MCTrk — Type

struct MCTrk

A simulated (Monte Carlo, hence "MC") track (or shower).

Fields

id::Int64

pos::Position{Float64}

dir::Direction{Float64}

t::Float64

E::Float64

len::Float64

type::Int32

status::Int32

mother_id::Int32

counter::Int32

KM3io.FitInformation — Type

struct FitInformation

A container object to store fit information which uses 0-based indexing. It implements the array interface. The entries is this vector are float values and their position encodes the meaning of their meaning which is defined in the KM3NeT DataFormat

Note

The elements of this object should always be accessed using constants defined in KM3io.FITINFORMATION. The use of magic numbers should be avoided to ensure that future changes in the KM3NeT DataFormat do not break existing code.

Fields

values::Vector{Float64}

Online Format

KM3io.DAQEvent — Type

struct DAQEvent

A (triggered) event holding snapshot hits and triggered hits. The triggered hits are a subset of the snapshot hits.

Fields

header::EventHeader

snapshot_hits::Vector{SnapshotHit}

triggered_hits::Vector{TriggeredHit}

KM3io.EventHeader — Type

struct EventHeader

The header of an event.

Fields

detector_id::Int32

run::Int32

frame_index::Int32

t::UTCExtended

trigger_counter::UInt64

trigger_mask::UInt64

overlays::UInt32

KM3io.SnapshotHit — Type

struct SnapshotHit <: KM3io.AbstractDAQHit

A snapshot hit.

Fields

dom_id::UInt32

channel_id::UInt8

t::Int32

tot::UInt8

KM3io.TriggeredHit — Type

struct TriggeredHit <: KM3io.AbstractDAQHit

A hit which was triggered.

Fields

dom_id::Int32

channel_id::UInt8

t::Int32

tot::UInt8

trigger_mask::UInt64

KM3io.CalibratedSnapshotHit — Type

struct CalibratedSnapshotHit <: AbstractCalibratedHit

A calibrated snapshot hit.

Fields

dom_id::UInt32

channel_id::UInt8

t::Float64

tot::UInt8

KM3io.CalibratedTriggeredHit — Type

struct CalibratedTriggeredHit <: AbstractCalibratedHit

A calibrated triggered hit.

Fields

dom_id::UInt32

channel_id::UInt8

t::Float64

tot::UInt8

trigger_mask::UInt64

KM3io.UTCTime — Type

struct UTCTime

A basic time structure with seconds and nanoseconds. The seconds are counting from the start of the epoch, just like the UNIX time.

Fields

s::UInt64

ns::UInt64

KM3io.UTCExtended — Type

struct UTCExtended

An extended time structure which contains the White Rabbit time synchronisation status. wr_status == 0 means that the synchronisation is OK.

Fields

s::UInt32

ns::UInt32

wr_status::Int64

KM3io.Summaryslice — Type

struct Summaryslice

A summaryslice is a condensed timeslice with the header information of the corresponding timeslice and a summary frame for each optical module. The hit information of the original timeslice is reduced so that for each PMT a single byte is used to encode the hit rate.

Fields

header::SummarysliceHeader

frames::Vector{SummaryFrame}

KM3io.SummarysliceHeader — Type

struct SummarysliceHeader

The header of a summaryslice.

Fields

detector_id::Int32

run::Int32

frame_index::Int32

t::UTCExtended

KM3io.SummaryFrame — Type

struct SummaryFrame

A SummaryFrame contains reduced timeslice data from an optical module.

The PMT rates are encoded as single bytes and can be converted to real hit rates using the rates(s::SummaryFrame) function.

Fields

dom_id::Int32

daq::UInt32

status::UInt32

fifo::UInt32

status3::UInt32

status4::UInt32

rates::StaticArraysCore.SVector{31, UInt8}

HDF5

KM3io.H5File — Type

struct H5File

A wrapper for an HDF5 file used in KM3NeT.

Fields

_h5f::HDF5.File

_datasets::Dict{String, H5CompoundDataset}

_lock::ReentrantLock

KM3io.H5CompoundDataset — Type

struct H5CompoundDataset{T}

A flat HDF5 compound dataset which is essentially a vector of structs. It has a cache which is filled when elements are pushed to it. The cache is automatically written to the target HDF5 path when full.

Fields

dset::HDF5.Dataset

cache::KM3io.H5CompoundDatasetCache

_lock::ReentrantLock

KM3io.addmeta — Function

addmeta(
    dset::Union{HDF5.Dataset, HDF5.Datatype, HDF5.File, HDF5.Group},
    object
)

Attaches key-value-pair meta entries to an HDF5 instance for each field of the given object.

HDF5.create_dataset — Function

create_dataset(
    f::H5File,
    path::AbstractString,
    ::Type{T};
    cache_size,
    chunk,
    filters,
    kwargs...
) -> H5CompoundDataset

Creates a one-dimensional compound dataset H5CompoundDataset of a given type which can be extended one-by-one. The cache is used to accumulate data and reduce the number of dataset extensions. Each time the cache is full, the HDF5 dataset will be extended, the buffer written and cleared.

To force the writing, use flush

Base.flush — Function

flush(d::H5CompoundDataset; nolock) -> H5CompoundDataset

Forces the cache to be written to the HDF5 file.

Hardware

KM3io.PMT — Type

struct PMT

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

dir::Direction{Float64}

t₀::Float64

status::Union{Missing, Int32}

KM3io.DetectorModule — Type

struct DetectorModule

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

location::Location

n_pmts::Int8

pmts::Vector{PMT}

q::Union{Missing, Quaternion{Float64}}

status::Int32

t₀::Float64

KM3io.Detector — Type

struct Detector

A KM3NeT detector.

Fields

version::Int8

id::Int32

validity::Union{Missing, KM3io.DateRange}

pos::Union{Missing, UTMPosition{Float64}}

utm_ref_grid::Union{Missing, String}

n_modules::Int32

modules::Dict{Int32, DetectorModule}

locations::Dict{Tuple{Int64, Int64}, DetectorModule}

strings::Vector{Int64}

comments::Vector{String}

KM3io.PMTFile — Type

struct PMTFile

A container type to hold PMT data which are stored in "PMT files", created by K40 calibrations. This type can be passe to Base.read to load the contents of such a file.

Example

julia> f = read("path/to/pmt.txt", PMTFile)
PMTFile containing parameters of 7254 PMTs

Fields

QE::Float64

mu::Float64

comments::Vector{String}

parameters::KM3io.PMTParameters

pmt_data::Dict{Tuple{Int64, Int64}, PMTData}

KM3io.PMTData — Type

struct PMTData

PMT parameters as stored in PMTFiles.

Fields

QE::Float64

gain::Float64

gainSpread::Float64

riseTime_ns::Float64

TTS_ns::Float64

threshold::Float64

KM3io.modules — Function

modules(d::Detector) -> Vector{DetectorModule}

Return a vector of all modules of a given detector.

KM3io.getmodule — Function

getmodule(d::Detector, module_id::Integer) -> DetectorModule

Return the detector module for a given module ID.

getmodule(
    d::Detector,
    string::Integer,
    floor::Integer
) -> DetectorModule

Return the detector module for a given string and floor.

getmodule(
    d::Detector,
    loc::Tuple{T<:Integer, T<:Integer}
) -> DetectorModule

Return the detector module for a given string and floor (as Tuple).

getmodule(d::Detector, loc::Location) -> DetectorModule

Return the detector module for a given location.

KM3io.getpmt — Function

getpmt(d::DetectorModule, channel_id::Integer) -> Any

Get the PMT for a given DAQ channel ID (TDC)

getpmt(d::Detector, hit) -> Any

Return the PMT for a given hit.

KM3io.haslocation — Function

haslocation(d::Detector, loc::Location) -> Bool

Returns true if there is a module at the given location.

KM3io.isbasemodule — Function

isbasemodule(d::DetectorModule) -> Bool

Returns true if the module is a basemodule.

KM3io.isopticalmodule — Function

isopticalmodule(d::DetectorModule) -> Bool

Returns true if the module is an optical module.

Base.write — Method

write(filename::AbstractString, d::Detector; version)

Writes the detector definition to a file, according to the DETX format specification. The version parameter can be a version number or :same, which is the default value and writes the same version as the provided detector has.

Base.write — Method

write(io::IO, d::Detector; version)

function write(io::IO, d::Detector; version=:same)

Writes the detector to a DETX formatted file. The target version can be specified via the version keyword. Note that if converting to higher versions, missing parameters will be filled with reasonable default values. In case of downgrading, information will be lost.

KM3io.Hydrophone — Type

struct Hydrophone

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

Fields

location::Location

pos::Position{Float64}

Base.read — Method

read(
    filename::AbstractString,
    T::Type{Hydrophone}
) -> Vector{Hydrophone}

function read(filename::AbstractString, T::Type{Hydrophone})

Reads a vector of Hydrophones from an ASCII file.

KM3io.Tripod — Type

struct Tripod

A tripod installed on the seabed which sends acoustic signals to modules.

Fields

id::Int8

pos::Position{Float64}

Base.read — Method

read(
    filename::AbstractString,
    T::Type{Tripod}
) -> Vector{Tripod}

function read(filename:AbstractString, T::Type{Tripod})

Reads a vector of Tripods from an ASCII file.

Base.write — Method

write(filename::AbstractString, tripods::Vector{Tripod})

function write(filename::AbstractString, tripods::Dict{Int8, Tripod})

Writes the position of tripods out into an ASCII file.

KM3io.piezoenabled — Function

piezoenabled(m::DetectorModule) -> Bool

function piezoenabled(m::DetectorModule)

Return true if the piezo is enabled, false otherwise.

KM3io.hydrophoneenabled — Function

hydrophoneenabled(m::DetectorModule) -> Bool

function hydrophonenabled(m::DetectorModule)

Return true if the hydrophone is enabled, false otherwise.

KM3io.center — Function

center(m::DetectorModule) -> Position{Float64}

Calculate the centre of a module by fitting the crossing point of the PMT axes.

center(d::Detector) -> Position{Float64}

Calculate the center of the detector based on the location of the optical modules.

KM3io.StringMechanics — Type

struct StringMechanics

A container structure which holds the mechanical model parameters for multiple strings, including a default value for strings which have specific parameters.

Fields

default::StringMechanicsParameters

stringparams::Dict{Int64, StringMechanicsParameters}

KM3io.StringMechanicsParameters — Type

struct StringMechanicsParameters

Data structure for parameters of the mechanical model of strings. This data structure is used to calculate the effective height conform to the mechanical model of the string.

Fields

a::Float64

b::Float64

Base.read — Method

read(
    filename::AbstractString,
    T::Type{StringMechanics}
) -> StringMechanics

Reads the mechanical models from a text file.

Optical Data

Acoustics

KM3io.Waveform — Type

struct Waveform

Waveform translates Emitter ID to Tripod ID.

Fields

ids::Dict{Int8, Int8}

Base.read — Method

read(
    filename::AbstractString,
    T::Type{Waveform}
) -> Waveform

function read(filename::AbstractString, T::Type{Waveform})

Reads the waveform ASCII file.

KM3io.AcousticSignal — Type

struct AcousticSignal

AcousticSignal is a custom type with four fields to store all the information inside the raw acoustic binary files.

dom_id::Int32 ID of the module

utc_seconds:: UInt32 storing the first 4 Bytes and is a UNIX time stamp

ns_cycles:: UInt32 storing the second 4 Bytes

samples:: UInt32 storing the third 4 Bytes, corresponding to the number of data points accuired during the measring window

pcm:: Vector of Float32 of length frame_length, storing all other 4 Byte blocks. Each entry is a data point of the acoustic signal.

Fields

dom_id::Int32

utc_seconds::UInt32

ns_cycles::UInt32

samples::UInt32

pcm::Vector{Float32}

KM3io.AcousticsTriggerParameter — Type

struct AcousticsTriggerParameter

Certain parameters which define an acoustic event.

Fields

q::Float64

tmax::Float64

nmin::Int32

Base.read — Method

read(
    filename::AbstractString,
    T::Type{AcousticsTriggerParameter}
) -> AcousticsTriggerParameter

function read(filename::AbstractString, T::Type{AcousticsTriggerParameter})

Reads the 'acousticstriggerparameters.txt' file.

Calibration

KM3io.calibrate — Function

calibrate(det::Detector, hits) -> Vector{XCalibratedHit}

Apply full geometry and time calibration to given hits. This way of calibration should be used wisely since it creates a very bloaded XCalibratedHit object, which might not be necessary. Often, we only need time the calibration to be applied.

KM3io.calibratetime — Function

calibratetime(
    det::Detector,
    hits::Array{T<:SnapshotHit, 1}
) -> Vector{CalibratedSnapshotHit}

Calibrate the time of a given array of snapshot hits.

calibratetime(
    det::Detector,
    hits::Array{T<:TriggeredHit, 1}
) -> Vector{CalibratedTriggeredHit}

Calibrate the time of a given array of triggered hits.

KM3io.combine — Function

combine(
    snapshot_hits::Vector{SnapshotHit},
    triggered_hits::Vector{TriggeredHit}
) -> Vector{TriggeredHit}

Combine snapshot and triggered hits to a single hits-vector.

This should be used to transfer the trigger information to the snapshot hits from a DAQEvent. The triggered hits are a subset of the snapshot hits.

KM3io.Orientations — Type

struct Orientations

A data structure to hold orientations data. This struct should be instantiated by Base.read(filename, Orientations).

Fields

module_ids::Set{Int64}

times::Dict{Int64, Vector{Float64}}

quaternions::Dict{Int64, Vector{Quaternion}}

KM3io.Compass — Type

struct Compass{T} <: StaticArraysCore.FieldVector{3, T}

A compass with yaw, pitch and roll.

Fields

yaw::Any

pitch::Any

roll::Any

KM3io.Quaternion — Type

struct Quaternion{T} <: StaticArraysCore.FieldVector{4, T}

A simple quaternion derived from a FieldVector of StaticArrays, no more, no less.

Fields

q0::Any

qx::Any

qy::Any

qz::Any

KM3io.floordist — Function

floordist(det::Detector) -> Float64

Calculates the average floor distance between neighboured modules.

KM3io.slew — Function

slew(tot::Integer) -> Any

Return the time slewing for a ToT.

KM3io.slerp — Function

slerp(q₁, q₂, t::Real; dot_threshold, normalized) -> Any

Interpolate between two vectors (e.g. quaternions) using the slerp method. t should be between 0 and 1. 0 will produce q₁ and 1 q₂.

The input vectors q₁ and q₂ will be normalised unless normalized is false. It is not done by default to shave off a few dozens of nanoseconds. Make sure to set normalized=false if the input vectors are not unit vectors.

Physics

KM3io.azimuth — Function

azimuth(d::Direction) -> Any

Calculates the azimuth from a given Direction according to KM3NeT conventions.

KM3io.zenith — Function

zenith(d::Direction) -> Any

Calculates the zenith from a given Direction according to KM3NeT conventions.

KM3io.phi — Function

phi(d::Direction) -> Any

Calculate the phi angle for a given direction

KM3io.theta — Function

theta(d::Direction) -> Any

Calculate the theta angle for a given direction

KM3io.cherenkov — Function

cherenkov(
    track,
    hits::Array{T<:AbstractCalibratedHit, 1}
) -> Vector

Calculates the parameters of cherenkov photons emitted from a track and hitting the PMTs represented as (calibrated) hits. The returned cherenkov photons hold information about the closest distance to track, the time residual, arrival time, impact angle, photon travel distance, track travel distance and photon travel direction. See CherenkovPhoton for more information.

KM3io.CherenkovPhoton — Type

struct CherenkovPhoton

A Cherenkov photon with parameters calculated from its inducing track. See cherenkov() for more information.

Fields

d_closest::Float64

d_photon::Float64

d_track::Float64

t::Float64

Δt::Float64

impact_angle::Float64

dir::Direction{Float64}

KM3io.K40Rates — Type

struct K40Rates

K40 rates with L0 and higher level rates (with increasing multiplicities).

Fields

L0::Float64

L1::Vector{Float64}

Trigger

KM3io.triggered — Function

triggered(e) -> Vector{T} where T<:KM3io.AbstractDAQHit

Return true if the passed object (hit, event, ...) was triggered by any trigger algorithm.

KM3io.is3dmuon — Function

is3dmuon(e::DAQEvent) -> Bool

is3dmuon(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:12.

Return true the 3D Muon trigger bit is set.

KM3io.is3dshower — Function

is3dshower(e::DAQEvent) -> Bool

is3dshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:20.

Return true if the 3D Shower trigger bit is set.

KM3io.ismxshower — Function

ismxshower(x) -> Bool

ismxshower(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:28.

ismxshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:29.

Return true if the MX Shower trigger bit is set.

KM3io.isnb — Function

isnb(x) -> Bool

isnb(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:36.

isnb(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:37.

Return true if the NanoBeacon trigger bit is set.

ControlHost

KM3io.CHClient — Type

struct CHClient{T}

A ControlHost client which can communicate with a Ligier dispatcher to receive messages for all the subscribed tags.

To connect to a Ligier which is receiving triggered DAQ events e.g. in the KM3NeT monitoring system or in a test setup consisting of a JLigier dispatcher and a JRegurgitate instance which is redispatching DAQ events (JDAQEvent) from a ROOT file in online format to the JLigier, a CHClient can be created to subscribe the event messages with

julia> using KM3io

julia> c = CHClient{DAQEvent}(ip"127.0.0.1", 5553)

julia> for event in c
           @show event
       end
e = KM3io.DAQEvent with 127 snapshot and 6 triggered hits
e = KM3io.DAQEvent with 147 snapshot and 6 triggered hits
e = KM3io.DAQEvent with 154 snapshot and 8 triggered hits
e = KM3io.DAQEvent with 152 snapshot and 6 triggered hits
...
...
...

Fields

ip::Sockets.IPv4

port::UInt16

tags::Vector{CHTag}

socket::Sockets.TCPSocket

Tools

General tools

KM3io.SummarysliceIntervalIterator — Type

struct SummarysliceIntervalIterator

An iterator which yields a Vector{Summaryslice} containing summaryslices of a given time_interval (in seconds). Useful when analysing summary data with fixed time intervals. The returned summaryslices are also sorted in time.

Examples

julia> f = ROOTFile("KM3NeT_00000133_00014728.root")
ROOTFile{OnlineTree (83509 events, 106969 summaryslices)}

julia> sii = SummarysliceIntervalIterator(f, 60)
SummarysliceIntervalIterator (10739.7s, 60s intervals, 179 chunks)

julia> for summaryslices in sii
           @show length(summaryslices)
           @show summaryslices[1].header
           break
       end
length(summaryslices) = 599
(summaryslices[1]).header = SummarysliceHeader(133, 14728, 134, UTCExtended(1676246413, 400000000, 0))

Note

Short time intervals (a few tens of seconds) will likely return Vector{Summaryslice}s with few entries in the first and last iterations due to a delay in run changes. The number of frames per summaryslice will gradually increase due to the asynchronous nature of the run transition. See the example below with a time inteval of 10s and 100 active optical modules.

julia> sii = SummarysliceIntervalIterator(f, 10)
SummarysliceIntervalIterator (106.2s, 10s intervals, 11 chunks)

julia> for summaryslices in sii
           n = length(summaryslices)
           @show n
       end
n = 73
n = 100
n = 100
n = 100
n = 100
n = 100
n = 100
n = 100
n = 100
n = 96
n = 31

Fields

sc::KM3io.SummarysliceContainer

first_frame_index::Int64

time_interval::Int64

n_chunks::Int64

timespan::Float64

indices::Vector{Int64}

KM3io.getevent — Function

getevent(
    tree::Union{KM3io.OnlineTree, KM3io.OfflineTree},
    frame_index,
    trigger_counter
) -> Any

Retrieves the event with for a given frame_index and trigger_counter.

KM3io.categorize — Function

categorize(
    field::Symbol,
    elements::Vector
) -> Dict{_A, Vector{_A1}} where {_A, _A1}

Categorise the struct elements of a vector by a given field into a dictionary of T.field => Vector{T}.

Examples

julia> using KM3io

julia> struct PMT  # just an ad-hoc PMT struct for demonstration purposes
         dom_id
         time
       end

julia> pmts = [PMT(2, 10.4), PMT(4, 23.5), PMT(2, 42.0)];

julia> categorize(:dom_id, pmts)
Dict{Any, Vector{PMT}} with 2 entries:
  4 => [PMT(4, 23.5)]
  2 => [PMT(2, 10.4), PMT(2, 42.0)]

KM3io.nthbitset — Function

nthbitset(n, a) -> Any

Return true if the n-th bit of a is set, false otherwise.

KM3io.most_frequent — Function

most_frequent(iterable) -> Any

Return the most frequent value of a given iterable.

most_frequent(f::Function, iterable; rettype) -> Int64

Return the most frequent value of a given iterable based on the return value of a function f which returns (hashable) values of rettype.

DAQ

KM3io.pmtrate — Function

pmtrate(r::UInt8) -> Float64

Calculate the PMT hit rate from the raw byte stored in a summary frame.

The rate of each PMT is encoded in a single byte to reduce the size of the summary frame, therefore only 256 values are possible which are mapped to an exponential function. The values are precalculated by the _getrate() function for the best performance.

KM3io.pmtrates — Function

pmtrates(
    s::SummaryFrame
) -> StaticArraysCore.SVector{31, Float64}

Return the actual rates (in Hz) for each PMT in a summary frame.

pmtrates(
    s::Summaryslice
) -> Dict{Int32, StaticArraysCore.SVector{31, Float64}}

Return a dictionary of DOM IDs as keys and PMT rates [Hz] as values (Vector{Float64}).

KM3io.hrvstatus — Function

hrvstatus(f::SummaryFrame, tdc) -> Any

Return true if the TDC is in high rate veto.

hrvstatus(f::SummaryFrame) -> Bool

Return true if any of the TDCs is in high rate veto.

KM3io.fifostatus — Function

fifostatus(f::SummaryFrame, tdc) -> Any

Return true if the TDC has FIFO almost full.

fifostatus(f::SummaryFrame) -> Bool

Return true if any of the TDCs is in high rate veto.

KM3io.tdcstatus — Function

tdcstatus(f::SummaryFrame) -> Bool

Return true if TDC status is OK.

KM3io.wrstatus — Function

wrstatus(f::SummaryFrame) -> Bool

Return true if White Rabbit status is OK.

KM3io.hasudptrailer — Function

hasudptrailer(f::SummaryFrame) -> Bool

Return true if the UDP trailer is present.

KM3io.count_active_channels — Function

count_active_channels(f::SummaryFrame) -> Int64

Number of TDCs without high rate veto or FIFO almost full.

KM3io.count_fifostatus — Function

count_fifostatus(f::SummaryFrame) -> Int64

Number of TDCs with FIFO almost full.

KM3io.count_hrvstatus — Function

count_hrvstatus(f::SummaryFrame) -> Int64

Number of TDCs with high rate veto.

KM3io.status — Function

status(f::SummaryFrame) -> Bool

Return true if TDC and White Rabbit status are OK.

KM3io.number_of_udp_packets_received — Function

number_of_udp_packets_received(f::SummaryFrame) -> Int32

Number of received UDP packets (excluding the trailer).

KM3io.maximal_udp_sequence_number — Function

maximal_udp_sequence_number(f::SummaryFrame) -> Int32

Maximal sequence number of all received UDP packets.

Reconstruction

KM3io.besttrack — Function

besttrack(
    e::Evt,
    rec_type::Integer,
    rsr::RecStageRange
) -> Union{Missing, Trk}

Return the best reconstructed track for a given reconstruction type and reconstruction stage range. If no track could be found, missing is returned.

KM3io.bestjppmuon — Function

bestjppmuon(e::Evt) -> Union{Missing, Trk}

Returns the best reconstructed JMuon track of an event or missing if there are none.

bestjppmuon(trks::Vector{Trk}) -> Union{Missing, Trk}

Returns the best reconstructed JMuon track or missing if there are none.

KM3io.bestjppshower — Function

bestjppshower(e::Evt) -> Union{Missing, Trk}

Returns the best reconstructed JShower "track" of an event or missing if there are none.

bestjppshower(trks::Vector{Trk}) -> Union{Missing, Trk}

Returns the best reconstructed JShower "track" or missing if there are none.

KM3io.bestaashower — Function

bestaashower(e::Evt) -> Union{Missing, Trk}

Returns the best reconstructed aashower "track" of an event or missing if there are none.

bestaashower(trks::Vector{Trk}) -> Union{Missing, Trk}

Returns the best reconstructed aashower "track" or missing if there are none.

KM3io.RecStageRange — Type

struct RecStageRange{T<:Integer}

This struct is used to represent a range of reconstruction stages. These are well-defined integers (see KM3NeT Dataformat) for each reconstruction algorithm and are stored in a vector named rec_stages of each Trk.

julia> using KM3io

julia> rsr = RecStageRange(KM3io.RECONSTRUCTION.JMUONBEGIN, KM3io.RECONSTRUCTION.JMUONEND)
RecStageRange{Int64}(0, 99)

julia> KM3io.RECONSTRUCTION.JMUONSIMPLEX ∈ rsr
true

julia> KM3io.RECONSTRUCTION.AASHOWERFITPREFIT ∈ rsr
false

julia> 23 ∈ rsr
true

julia> 523 ∈ rsr
false

Fields

lower::Integer

upper::Integer

KM3io.hashistory — Function

hashistory(
    t::Trk,
    rec_type::Integer,
    rsr::RecStageRange
) -> Bool

Returns true if a track with a given rec_type contains all the reconstruction stages in rsr::RecStageRange.

hashistory(
    t::Trk,
    rec_type::Integer,
    rec_stage::Integer
) -> Bool

Returns true if a track with a given rec_type contains the rec_stage.

Math

Base.angle — Function

angle(d1::Direction, d2::Direction) -> Any

Calculate the angle between two vectors.

KM3io.distance — Function

distance(a::Position, b::Position) -> Any

Calculates the disance between two points.