API

KM3io.AcousticSignal

KM3io.AcousticsTriggerParameter

KM3io.Detector

KM3io.DetectorModule

KM3io.Direction

KM3io.Hydrophone

KM3io.Location

KM3io.PMT

KM3io.Position

KM3io.RecStageRange

KM3io.Tripod

KM3io.Trk

KM3io.Waveform

KM3io.besttrack

KM3io.calibrate

KM3io.categorize

KM3io.center

KM3io.floordist

KM3io.hashistory

KM3io.hydrophoneenabled

KM3io.is3dmuon

KM3io.is3dshower

KM3io.ismxshower

KM3io.isnb

KM3io.most_frequent

KM3io.nthbitset

KM3io.piezoenabled

KM3io.slew

Basic Data Structures

KM3io.Position — Type

struct Position{T} <: StaticArraysCore.FieldVector{3, T}

A vector to represent a position in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Direction — Type

struct Direction{T} <: StaticArraysCore.FieldVector{3, T}

A vector to represent a direction in 3D.

Fields

x::Any

y::Any

z::Any

KM3io.Location — Type

struct Location

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

Fields

string::Int32

floor::Int8

Offline Format

KM3io.Trk — Type

struct Trk

Represents a reconstructed "track", which can be e.g. a muon track but also a shower.

Fields

id::Int64

pos::Position{Float64}

dir::Direction{Float64}

t::Float64

E::Float64

len::Float64

lik::Float64

rec_type::Int32

rec_stages::Vector{Int32}

fitinf::Vector{Float64}

Missing docstring.

Missing docstring for McTrk. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Evt. Check Documenter's build log for details.

Online Format

Hardware

KM3io.PMT — Type

struct PMT

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

dir::Direction{Float64}

t₀::Float64

status::Union{Missing, Int32}

KM3io.DetectorModule — Type

struct DetectorModule

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

Fields

id::Int32

pos::Position{Float64}

location::Location

n_pmts::Int8

pmts::Vector{PMT}

q::Union{Missing, Quaternion{Float64}}

status::Int32

t₀::Float64

KM3io.Detector — Type

struct Detector

A KM3NeT detector.

Fields

version::Int8

id::Int32

validity::Union{Missing, KM3io.DateRange}

pos::Union{Missing, UTMPosition{Float64}}

utm_ref_grid::Union{Missing, String}

n_modules::Int32

modules::Dict{Int32, DetectorModule}

locations::Dict{Tuple{Int64, Int64}, DetectorModule}

strings::Vector{Int64}

comments::Vector{String}

KM3io.Hydrophone — Type

struct Hydrophone

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

Fields

location::Location

pos::Position{Float64}

KM3io.Tripod — Type

struct Tripod

A tripod installed on the seabed which sends acoustic signals to modules.

Fields

id::Int8

pos::Position{Float64}

KM3io.piezoenabled — Function

piezoenabled(m::DetectorModule) -> Bool

function piezoenabled(m::DetectorModule)

Return true if the piezo is enabled, false otherwise.

KM3io.hydrophoneenabled — Function

hydrophoneenabled(m::DetectorModule) -> Bool

function hydrophonenabled(m::DetectorModule)

Return true if the hydrophone is enabled, false otherwise.

KM3io.center — Function

center(m::DetectorModule) -> Position{Float64}

Calculate the centre of a module by fitting the crossing point of the PMT axes.

Optical Data

Acoustics

KM3io.Waveform — Type

struct Waveform

Waveform translates Emitter ID to Tripod ID.

Fields

ids::Dict{Int8, Int8}

KM3io.AcousticSignal — Type

struct AcousticSignal

AcousticSignal is a custom type with four fields to store all the information inside the raw acoustic binary files.

dom_id::Int32 ID of the module

utc_seconds:: UInt32 storing the first 4 Bytes and is a UNIX time stamp

ns_cycles:: UInt32 storing the second 4 Bytes

samples:: UInt32 storing the third 4 Bytes, corresponding to the number of data points accuired during the measring window

pcm:: Vector of Float32 of length frame_length, storing all other 4 Byte blocks. Each entry is a data point of the acoustic signal.

Fields

dom_id::Int32

utc_seconds::UInt32

ns_cycles::UInt32

samples::UInt32

pcm::Vector{Float32}

KM3io.AcousticsTriggerParameter — Type

struct AcousticsTriggerParameter

Certain parameters which define an acoustic event.

Fields

q::Float64

tmax::Float64

nmin::Int32

Calibration

KM3io.calibrate — Function

calibrate(
    det::Detector,
    hits
) -> Vector{KM3io.CalibratedDAQHit}

Apply geometry and time calibration to given hits.

KM3io.floordist — Function

floordist(det::Detector) -> Float64

Calculates the average floor distance between neighboured modules.

KM3io.slew — Function

slew(tot::Integer) -> Any

Return the time slewing for a ToT.

Trigger

KM3io.is3dmuon — Function

is3dmuon(e::KM3io.DAQEvent) -> Bool

is3dmuon(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:11.

Return true the 3D Muon trigger bit is set.

KM3io.is3dshower — Function

is3dshower(e::KM3io.DAQEvent) -> Bool

is3dshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:19.

Return true if the 3D Shower trigger bit is set.

KM3io.ismxshower — Function

ismxshower(x) -> Bool

ismxshower(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:27.

ismxshower(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:28.

Return true if the MX Shower trigger bit is set.

KM3io.isnb — Function

isnb(x) -> Bool

isnb(x)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:35.

isnb(e)

defined at /builds/common/KM3io.jl/src/tools/trigger.jl:36.

Return true if the NanoBeacon trigger bit is set.

Tools

General tools

KM3io.categorize — Function

categorize(field::Symbol, elements::Vector) -> Dict

Categorise the struct elements of a vector by a given field into a dictionary of T.field => Vector{T}.

Examples

julia> using KM3io

julia> struct PMT  # just an ad-hoc PMT struct for demonstration purposes
         dom_id
         time
       end

julia> pmts = [PMT(2, 10.4), PMT(4, 23.5), PMT(2, 42.0)];

julia> categorize(:dom_id, pmts)
Dict{Any, Vector{PMT}} with 2 entries:
  4 => [PMT(4, 23.5)]
  2 => [PMT(2, 10.4), PMT(2, 42.0)]

KM3io.nthbitset — Function

nthbitset(n, a) -> Any

Return true if the n-th bit of a is set, false otherwise.

KM3io.most_frequent — Function

most_frequent(iterable) -> Any

Return the most frequent value of a given iterable.

most_frequent(f::Function, iterable; rettype) -> Int64

Return the most frequent value of a given iterable based on the return value of a function f which returns (hashable) values of rettype.

Reconstruction

KM3io.besttrack — Function

besttrack(
    e::KM3io.Evt,
    rec_type::Integer,
    rsr::RecStageRange
) -> Union{Nothing, Trk}

Return the best reconstructed track for a given reconstruction type and reconstruction stage range. If no track could be found, nothing is returned.

Missing docstring.

Missing docstring for bestjppmuon. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bestjppshower. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bestaashower. Check Documenter's build log for details.

KM3io.RecStageRange — Type

struct RecStageRange{T<:Integer}

This struct is used to represent a range of reconstruction stages. These are well-defined integers (see KM3NeT Dataformat) for each reconstruction algorithm and are stored in a vector named rec_stages of each Trk.

julia> using KM3io

julia> rsr = RecStageRange(KM3io.RECONSTRUCTION.JMUONBEGIN, KM3io.RECONSTRUCTION.JMUONEND)
RecStageRange{Int64}(0, 99)

julia> KM3io.RECONSTRUCTION.JMUONSIMPLEX ∈ rsr
true

julia> KM3io.RECONSTRUCTION.AASHOWERFITPREFIT ∈ rsr
false

julia> 23 ∈ rsr
true

julia> 523 ∈ rsr
false

Fields

lower::Integer

upper::Integer

KM3io.hashistory — Function

hashistory(
    t::Trk,
    rec_type::Integer,
    rsr::RecStageRange
) -> Bool

Returns true if a track with a given rec_type contains all the reconstruction stages in rsr::RecStageRange.

hashistory(
    t::Trk,
    rec_type::Integer,
    rec_stage::Integer
) -> Union{Missing, Bool}

Returns true if a track with a given rec_type contains the rec_stage.